+++ integration/fieldoption-jstype-with-forcelong-bigint/fieldoption-jstype-with-forcelong-bigint.ts
@@ -13,12 +13,12 @@ export interface FieldOption {
 }
 
 function createBaseFieldOption(): FieldOption {
-  return { normalField: BigInt("0"), numberField: 0, stringField: "0" };
+  return { normalField: 0n, numberField: 0, stringField: "0" };
 }
 
 export const FieldOption = {
   encode(message: FieldOption, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
-    if (message.normalField !== BigInt("0")) {
+    if (message.normalField !== 0n) {
       if (BigInt.asIntN(64, message.normalField) !== message.normalField) {
         throw new globalThis.Error("value provided for field message.normalField of type int64 too large");
       }
@@ -78,7 +78,7 @@ export const FieldOption = {
 
   fromJSON(object: any): FieldOption {
     return {
-      normalField: isSet(object.normalField) ? BigInt(object.normalField) : BigInt("0"),
+      normalField: isSet(object.normalField) ? BigInt(object.normalField) : 0n,
       numberField: isSet(object.numberField) ? globalThis.Number(object.numberField) : 0,
       stringField: isSet(object.stringField) ? globalThis.String(object.stringField) : "0",
     };
@@ -86,7 +86,7 @@ export const FieldOption = {
 
   toJSON(message: FieldOption): unknown {
     const obj: any = {};
-    if (message.normalField !== BigInt("0")) {
+    if (message.normalField !== 0n) {
       obj.normalField = message.normalField.toString();
     }
     if (message.numberField !== 0) {
@@ -103,7 +103,7 @@ export const FieldOption = {
   },
   fromPartial<I extends Exact<DeepPartial<FieldOption>, I>>(object: I): FieldOption {
     const message = createBaseFieldOption();
-    message.normalField = object.normalField ?? BigInt("0");
+    message.normalField = object.normalField ?? 0n;
     message.numberField = object.numberField ?? 0;
     message.stringField = object.stringField ?? "0";
     return message;