+++ integration/simple-long-bigint-noliteral/google/protobuf/wrappers.ts
@@ -211,12 +211,12 @@ export const FloatValue = {
 };
 
 function createBaseInt64Value(): Int64Value {
-  return { value: 0n };
+  return { value: BigInt("0") };
 }
 
 export const Int64Value = {
   encode(message: Int64Value, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
-    if (message.value !== 0n) {
+    if (message.value !== BigInt("0")) {
       if (BigInt.asIntN(64, message.value) !== message.value) {
         throw new globalThis.Error("value provided for field message.value of type int64 too large");
       }
@@ -249,12 +249,12 @@ export const Int64Value = {
   },
 
   fromJSON(object: any): Int64Value {
-    return { value: isSet(object.value) ? BigInt(object.value) : 0n };
+    return { value: isSet(object.value) ? BigInt(object.value) : BigInt("0") };
   },
 
   toJSON(message: Int64Value): unknown {
     const obj: any = {};
-    if (message.value !== 0n) {
+    if (message.value !== BigInt("0")) {
       obj.value = message.value.toString();
     }
     return obj;
@@ -265,18 +265,18 @@ export const Int64Value = {
   },
   fromPartial<I extends Exact<DeepPartial<Int64Value>, I>>(object: I): Int64Value {
     const message = createBaseInt64Value();
-    message.value = object.value ?? 0n;
+    message.value = object.value ?? BigInt("0");
     return message;
   },
 };
 
 function createBaseUInt64Value(): UInt64Value {
-  return { value: 0n };
+  return { value: BigInt("0") };
 }
 
 export const UInt64Value = {
   encode(message: UInt64Value, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
-    if (message.value !== 0n) {
+    if (message.value !== BigInt("0")) {
       if (BigInt.asUintN(64, message.value) !== message.value) {
         throw new globalThis.Error("value provided for field message.value of type uint64 too large");
       }
@@ -309,12 +309,12 @@ export const UInt64Value = {
   },
 
   fromJSON(object: any): UInt64Value {
-    return { value: isSet(object.value) ? BigInt(object.value) : 0n };
+    return { value: isSet(object.value) ? BigInt(object.value) : BigInt("0") };
   },
 
   toJSON(message: UInt64Value): unknown {
     const obj: any = {};
-    if (message.value !== 0n) {
+    if (message.value !== BigInt("0")) {
       obj.value = message.value.toString();
     }
     return obj;
@@ -325,7 +325,7 @@ export const UInt64Value = {
   },
   fromPartial<I extends Exact<DeepPartial<UInt64Value>, I>>(object: I): UInt64Value {
     const message = createBaseUInt64Value();
-    message.value = object.value ?? 0n;
+    message.value = object.value ?? BigInt("0");
     return message;
   },
 };