+++ src/main.ts
@@ -101,6 +101,7 @@ import {
   maybeAddComment,
   maybePrefixPackage,
   nullOrUndefined,
+  oneofValueName,
   safeAccessor,
   withAndMaybeCheckIsNotNull,
   withOrMaybeCheckIsNotNull,
@@ -611,6 +612,11 @@ function makeDeepPartial(options: Options, longs: ReturnType<typeof makeLongUtil
       : T extends { ${maybeReadonly(options)}$case: string }
       ? { [K in keyof Omit<T, '$case'>]?: DeepPartial<T[K]> } & { ${maybeReadonly(options)}$case: T['$case'] }
     `;
+  } else if (options.oneof === OneofOption.UNIONS_VALUE) {
+    oneofCase = `
+      : T extends { ${maybeReadonly(options)}$case: string; value: unknown; }
+      ? { ${maybeReadonly(options)}$case: T['$case']; value?: DeepPartial<T['value']>; }
+    `;
   }
 
   const maybeExport = options.exportCommonSymbols ? "export" : "";
@@ -974,7 +980,8 @@ function generateOneofProperty(
     fields.map((f) => {
       let fieldName = maybeSnakeToCamel(f.name, options);
       let typeName = toTypeName(ctx, messageDesc, f);
-      return code`{ ${mbReadonly}$case: '${fieldName}', ${mbReadonly}${fieldName}: ${typeName} }`;
+      let valueName = oneofValueName(fieldName, options);
+      return code`{ ${mbReadonly}$case: '${fieldName}', ${mbReadonly}${valueName}: ${typeName} }`;
     }),
     { on: " | " },
   );
@@ -1283,9 +1290,10 @@ function generateDecode(ctx: Context, fullName: string, messageDesc: DescriptorP
       const oneofNameWithMessage = options.useJsonName
         ? messageProperty
         : getPropertyAccessor("message", maybeSnakeToCamel(messageDesc.oneofDecl[field.oneofIndex].name, options));
+      const valueName = oneofValueName(fieldName, options);
       chunks.push(code`
         ${tagCheck}
-        ${oneofNameWithMessage} = { $case: '${fieldName}', ${fieldName}: ${readSnippet} };
+        ${oneofNameWithMessage} = { $case: '${fieldName}', ${valueName}: ${readSnippet} };
       `);
     } else {
       chunks.push(code`
@@ -1590,8 +1598,9 @@ function generateEncode(ctx: Context, fullName: string, messageDesc: DescriptorP
         for (const oneOfField of oneOfFieldsDict[field.oneofIndex]) {
           const writeSnippet = getEncodeWriteSnippet(ctx, oneOfField);
           const oneOfFieldName = maybeSnakeToCamel(oneOfField.name, ctx.options);
+          const valueName = oneofValueName(oneOfFieldName, ctx.options);
           chunks.push(code`case "${oneOfFieldName}":
-            ${writeSnippet(`${oneofNameWithMessage}.${oneOfFieldName}`)};
+            ${writeSnippet(`${oneofNameWithMessage}.${valueName}`)};
             break;`);
         }
         chunks.push(code`}`);
@@ -2077,15 +2086,17 @@ function generateFromJson(ctx: Context, fullName: string, fullTypeName: string, 
       } else {
         const fallback = noDefaultValue ? nullOrUndefined(options) : "[]";
 
-        const readValueSnippet = readSnippet("e");
-        if (readValueSnippet.toString() === code`e`.toString()) {
+        const needMap = readSnippet("e").toString() !== code`e`.toString();
+        if (!needMap) {
           chunks.push(
             code`${fieldKey}: ${ctx.utils.globalThis}.Array.isArray(${jsonPropertyOptional}) ? [...${jsonProperty}] : [],`,
           );
         } else {
           // Explicit `any` type required to make TS with noImplicitAny happy. `object` is also `any` here.
           chunks.push(code`
-            ${fieldKey}: ${ctx.utils.globalThis}.Array.isArray(${jsonPropertyOptional}) ? ${jsonProperty}.map((e: any) => ${readValueSnippet}): ${fallback},
+            ${fieldKey}: ${
+            ctx.utils.globalThis
+          }.Array.isArray(${jsonPropertyOptional}) ? ${jsonProperty}.map((e: any) => ${readSnippet("e")}): ${fallback},
           `);
         }
       }
@@ -2099,8 +2110,9 @@ function generateFromJson(ctx: Context, fullName: string, fullTypeName: string, 
         chunks.push(code`${fieldName}: `);
       }
 
+      const valueName = oneofValueName(fieldKey, options);
       const ternaryIf = code`${ctx.utils.isSet}(${jsonProperty})`;
-      const ternaryThen = code`{ $case: '${fieldName}', ${fieldKey}: ${readSnippet(`${jsonProperty}`)}`;
+      const ternaryThen = code`{ $case: '${fieldName}', ${valueName}: ${readSnippet(`${jsonProperty}`)}`;
       chunks.push(code`${ternaryIf} ? ${ternaryThen}} : `);
 
       if (field === lastCase) {
@@ -2294,8 +2306,8 @@ function generateToJson(
       }
     } else if (isRepeated(field)) {
       // Arrays might need their elements transformed
-      const transformElement = readSnippet("e");
-      const maybeMap = transformElement.toCodeString([]) !== "e" ? code`.map(e => ${transformElement})` : "";
+      const needMap = readSnippet("e").toCodeString([]) !== "e";
+      const maybeMap = needMap ? code`.map(e => ${readSnippet("e")})` : "";
       chunks.push(code`
         if (${messageProperty}?.length) {
           ${jsonProperty} = ${messageProperty}${maybeMap};
@@ -2306,9 +2318,10 @@ function generateToJson(
       const oneofNameWithMessage = options.useJsonName
         ? messageProperty
         : getPropertyAccessor("message", maybeSnakeToCamel(messageDesc.oneofDecl[field.oneofIndex].name, options));
+      const valueName = oneofValueName(fieldName, options);
       chunks.push(code`
         if (${oneofNameWithMessage}?.$case === '${fieldName}') {
-          ${jsonProperty} = ${readSnippet(`${oneofNameWithMessage}.${fieldName}`)};
+          ${jsonProperty} = ${readSnippet(`${oneofNameWithMessage}.${valueName}`)};
         }
       `);
     } else {
@@ -2483,14 +2496,15 @@ function generateFromPartial(ctx: Context, fullName: string, messageDesc: Descri
       const oneofName = maybeSnakeToCamel(messageDesc.oneofDecl[field.oneofIndex].name, options);
       const oneofNameWithMessage = getPropertyAccessor("message", oneofName);
       const oneofNameWithObject = getPropertyAccessor("object", oneofName);
-      const v = readSnippet(`${oneofNameWithObject}.${fieldName}`);
+      const valueName = oneofValueName(fieldName, options);
+      const v = readSnippet(`${oneofNameWithObject}.${valueName}`);
       chunks.push(code`
         if (
           ${oneofNameWithObject}?.$case === '${fieldName}'
-          && ${oneofNameWithObject}?.${fieldName} !== undefined
-          && ${oneofNameWithObject}?.${fieldName} !== null
+          && ${oneofNameWithObject}?.${valueName} !== undefined
+          && ${oneofNameWithObject}?.${valueName} !== null
         ) {
-          ${oneofNameWithMessage} = { $case: '${fieldName}', ${fieldName}: ${v} };
+          ${oneofNameWithMessage} = { $case: '${fieldName}', ${valueName}: ${v} };
         }
       `);
     } else if (readSnippet(`x`).toCodeString([]) == "x") {